requirements.txt
argon2-cffi==20.1.0
asgiref==3.4.1
autopep8==1.4.4
bcrypt==3.1.7
beautifulsoup4==4.10.0
certifi==2021.5.30
cffi==1.14.6
charset-normalizer==2.0.6
colorama==0.4.4
cryptography==36.0.1
Django==3.2.7
django-bootstrap4==1.1.1
django-countries==7.2.1
djangorestframework==3.11.0
idna==3.2
pycodestyle==2.7.0
pycparser==2.20
pytz==2021.1
requests==2.26.0
six==1.16.0
soupsieve==2.2.1
sqlparse==0.4.2
urllib3==1.26.7
